User claims valid space and move is displayed
User input is validated so that a space must exist on the board and be
open, as well as rejecting input that isn't a number.

The state struct uses getter/setter methods so that other modules have an
abstraction to depend on rather than directly accessing the fields
within the state truct. Now if the fields within struct change only the
getter/setter methods need to change.

By adding an IO wrapper the "Main" function now can be passed an
environment variable in the form of an atom that dictates the output
device.