Define conditions for end of game
Status module generates every possible combination of moves that could
produce a win and uses those combinations to check if the status of the
game is over, tied, or won. The winning combinations are all built upon
first chunking the rows of the board. The rows are created by chunking
the board every n elements, where n is the square root of the size of
the board.