Rename Board to ConsoleBoardPresenter
Board was formerly concerned with both the storage mechanism of moves
and how the board is rendered and displayed in the console. Board has
been renamed to ConsoleBoardPresenter and is strictly focused on
rendering and displaying the board in the console.

The Functions that relate to the storage mechanism for moves have been
relocated to State module.