Add option for user to select player token
When game state is being initialized user is prompted to enter either an
X or an O as their player token. The token will be accepted regardless
of case as long as it's X or O. If it's not one of those two characters
the user will be prompted again to enter their desired symbol.