Add feature to choose game mode
When the game starts the user is presented with a choice to either play
against another human or against the computer player. Based on their
selection, an atom with the game mode is returned and passed into
Game.play(state, deps, mode). Game.play uses pattern matching based on
the mode to switch between the two game types.