User can choose alphanumeric player token
The user can now choose any alphanumerical character for their player
token. Status calculates a winning combination if all the elements in
the combination are the same and doesn't care about what the actual
character is.

This change necessitated adding two new fields to the state struct:
opponent and current player because there are no default player tokens,
and thus switching between players requires distinct struct fields. This
also adds the ability for the game to know who's turn it is currently
and block attempts to place a move by a user that isn't the current
player.

The only character not allowed is if the opponent's token is identical
to the player's token.

If it is the same the user will be prompted again to enter a unique
token.