Add unbeatable computer player
The unbeatable computer implements the minimax algorithm to ensure that
the worst outcome for the computer player is a tie. It also handles
cases where the player and opponent symbols are numbers that are also
positions on the board, since the symbols are represented as strings and
spots on the board are integers.