Require player token to be one character long
Validator checks to see if the token the user enters is one byte long,
and asks for another token if the given token is less or more than one
byte long. Checking for length of the player token in this way removes
capability to have emojis as player tokens since they're 4 bytes long. A
custom guard to check for the length of a string, not the number of
bytes, would allow for emojis to be entered.