User can select who goes first
User is now prompted to select which player goes first. They must enter
the exact player/opponent token to proceed, otherwise they're shown an
error message and are prompted again to choose which player they would
like to go first. Their choice is then set in the current_player field
in the State struct.

The functionality is implemented in Game.play, where the turn is
determined by who the current player is as determined by querying
state.current_player.